refactor: blacklist 출력 CustomerProgram 으로 변경
